fixed every issue the navbar had excluding the mobile nav not closing when a dropdown is active
