+++ frontend/directives/clickOutside.js
@@ -2,15 +2,13 @@
 export default {
   beforeMount(el, binding) {
     el.clickOutsideEvent = (event) => {
-      if (!(el === event.target || el.contains(event.target))) {
-        binding.value(event)
+      if (!el.contains(event.target) && !event.target.closest('[role="menu"]')) {
+        binding.value()
       }
     }
     document.addEventListener('click', el.clickOutsideEvent)
-    document.addEventListener('touchstart', el.clickOutsideEvent)
   },
   unmounted(el) {
     document.removeEventListener('click', el.clickOutsideEvent)
-    document.removeEventListener('touchstart', el.clickOutsideEvent)
   }
 }
